Adjust vgs implementation to better reflect how package will be used in checkout
