+++ Sources/ForageSDK/Foundation/ForageSDK+ForageServices.swift
@@ -43,6 +43,7 @@ protocol ForageSDKService: AnyObject {
         merchantAccount: String,
         paymentMethodReference: String,
         cardNumberToken: String,
+        foragePinTextEdit: ForagePINTextField,
         completion: @escaping (Result<Data?, Error>) -> Void)
     
     /// Capture a payment for a given payment reference
@@ -58,6 +59,7 @@ protocol ForageSDKService: AnyObject {
         merchantAccount: String,
         paymentReference: String,
         cardNumberToken: String,
+        foragePinTextEdit: ForagePINTextField,
         completion: @escaping (Result<Data?, Error>) -> Void)
     
     /// Cancel any ongoing request
@@ -100,6 +102,7 @@ extension ForageSDK: ForageSDKService {
         merchantAccount: String,
         paymentMethodReference: String,
         cardNumberToken: String,
+        foragePinTextEdit: ForagePINTextField,
         completion: @escaping (Result<Data?, Error>) -> Void) {
         service?.getXKey(bearerToken: bearerToken) { result in
             switch result {
@@ -112,7 +115,7 @@ extension ForageSDK: ForageSDKService {
                     merchantID: merchantAccount,
                     xKey: model.alias
                 )
-                self.service?.getBalance(request: request, completion: completion)
+                self.service?.getBalance(vgs: foragePinTextEdit.collector, request: request, completion: completion)
             case .failure(let error):
                 completion(.failure(error))
             }
@@ -124,6 +127,7 @@ extension ForageSDK: ForageSDKService {
         merchantAccount: String,
         paymentReference: String,
         cardNumberToken: String,
+        foragePinTextEdit: ForagePINTextField,
         completion: @escaping (Result<Data?, Error>) -> Void) {
         service?.getXKey(bearerToken: bearerToken) { result in
             switch result {
@@ -137,7 +141,7 @@ extension ForageSDK: ForageSDKService {
                     xKey: model.alias
                 )
                 
-                self.service?.requestCapturePayment(request: request, completion: completion)
+                self.service?.requestCapturePayment(vgs: foragePinTextEdit.collector, request: request, completion: completion)
             case .failure(let error):
                 completion(.failure(error))
             }
