+++ Sources/ForageSDK/Foundation/Network/LiveForageService.swift
@@ -9,17 +9,14 @@ import Foundation
 import VGSCollectSDK
 
 internal class LiveForageService: ForageService {
-    
     // MARK: Properties
     
     internal var provider: Provider
-    private var collector: VGSCollect?
     private var maxAttempts: Int = 10
     private var intervalBetweenAttempts: Double = 1.0
     private var retryCount = 0
     
-    init(_ collector: VGSCollect?, provider: Provider = Provider()) {
-        self.collector = collector
+    init(provider: Provider = Provider()) {
         self.provider = provider
     }
     
@@ -61,10 +58,11 @@ internal class LiveForageService: ForageService {
     ///  - request: Model element with data to perform request.
     ///  - completion: Returns balance object.
     internal func getBalance(
+        vgs: VGSCollect,
         request: ForageRequestModel,
         completion: @escaping (Result<Data?, Error>) -> Void) -> Void
     {
-        collector?.customHeaders = [
+        vgs.customHeaders = [
             "X-KEY": request.xKey,
             "IDEMPOTENCY-KEY": UUID.init().uuidString,
             "Merchant-Account": request.merchantID
@@ -74,7 +72,7 @@ internal class LiveForageService: ForageService {
             "card_number_token": request.cardNumberToken
         ]
 
-        collector?.sendData(
+        vgs.sendData(
             path: "/api/payment_methods/\(request.paymentMethodReference)/balance/",
             extraData: extraData) { [weak self] result in
                 self?.polling(response: result, request: request, completion: { pollingResult in
@@ -113,10 +111,11 @@ internal class LiveForageService: ForageService {
     ///  - request: Model element with data to perform request.
     ///  - completion: Returns captured payment object.
     internal func requestCapturePayment(
+        vgs: VGSCollect,
         request: ForageRequestModel,
         completion: @escaping (Result<Data?, Error>) -> Void)
     {
-        collector?.customHeaders = [
+        vgs.customHeaders = [
             "X-KEY": request.xKey,
             "IDEMPOTENCY-KEY": request.paymentReference,
             "Merchant-Account": request.merchantID
@@ -126,7 +125,7 @@ internal class LiveForageService: ForageService {
             "card_number_token": request.cardNumberToken
         ]
 
-        collector?.sendData(
+        vgs.sendData(
             path: "/api/payments/\(request.paymentReference)/capture/",
             extraData: extraData) { [weak self] result in
                 self?.polling(response: result, request: request, completion: { pollingResult in
